// front/src/app/api-client/model/routine-step-request.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Étape d\'une routine. Avec un id : étape existante (ses coches du jour sont gardées) ; sans : nouvelle étape.
 */
export interface RoutineStepRequest { 
    id?: number | null;
    libelle: string;
    icone: string;
    /**
     * Couleur de la ligne #RRGGBB ; absente = couleur pastel suivante
     */
    couleur?: string | null;
}